server/requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.10

# Environment Configuration
python-decouple==3.8

# Utilities
Pillow==11.0.0
python-dateutil==2.8.2

# Development (optional)
django-debug-toolbar==4.2.0
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Production (optional)
gunicorn==21.2.0
whitenoise==6.6.0
